% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angle_model}
\alias{angle_model}
\title{Angle Model}
\format{
A multinomial model
}
\usage{
angle_model
}
\description{
A model that predicts the number of Angle routes on a given play
}
\keyword{datasets}
