% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{block_model}
\alias{block_model}
\title{Block Snaps Model}
\format{
A linear model
}
\usage{
block_model
}
\description{
A model that predicts an OLineman's snaps for a given week
}
\keyword{datasets}
