% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockdf}
\alias{blockdf}
\title{PFF Blocking Snaps Data 2024}
\format{
A data frame with 339 rows and 3 variables:
\describe{
\item{player}{An NFL player}
\item{position}{All are OL (abbreviation of Offensive Line)}
\item{mean_snaps}{The mean number of snaps for the player in 2024}
}
}
\usage{
blockdf
}
\description{
A data set containing NFL offensive line player snap information for 2024.
}
\keyword{datasets}
