% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checkblock}
\alias{checkblock}
\title{OLine Depth Snaps}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{depth}{Position and spot on the depth chart}
\item{count}{Number of observations (players snaps in a given week)}
\item{mean_snaps}{The mean number of blocking snaps for that position + depth}
}
}
\usage{
checkblock
}
\description{
A data set containing average blocking snaps for OLinemen (OL) based on their
position in the depth chart.
}
\keyword{datasets}
