% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checkpassrush}
\alias{checkpassrush}
\title{Pass Rush Depth Snaps}
\format{
A data frame with 3 rows and 3 variables:
\describe{
\item{depth}{Position and spot on the depth chart}
\item{count}{Number of observations (players snaps in a given week)}
\item{mean_snaps}{The mean number of pass rush snaps for that position + depth}
}
}
\usage{
checkpassrush
}
\description{
A data set containing average pass rush snaps for front seven (F7) and secondary
(DB) players based on their position in the depth chart.
}
\keyword{datasets}
