% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checkrec}
\alias{checkrec}
\title{Receiver Depth Routes}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{depth}{Position and spot on the depth chart}
\item{count}{Number of observations (players snaps in a given week)}
\item{mean_snaps}{The mean number of routes ran for that position + depth}
}
}
\usage{
checkrec
}
\description{
A data set containing average routes ran for wide receivers (WR) and tight
ends (TE) based on their position in the depth chart.
}
\keyword{datasets}
