% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{checkrundef}
\alias{checkrundef}
\title{Run Defense Depth Snaps}
\format{
A data frame with 6 rows and 3 variables:
\describe{
\item{depth}{Position and spot on the depth chart}
\item{count}{Number of observations (players snaps in a given week)}
\item{mean_snaps}{The mean number of run defense snaps for that position + depth}
}
}
\usage{
checkrundef
}
\description{
A data set containing average run defense snaps for front seven (F7) and secondary
(DB) players based on their position in the depth chart.
}
\keyword{datasets}
