% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corner_model}
\alias{corner_model}
\title{Corner Model}
\format{
A multinomial model
}
\usage{
corner_model
}
\description{
A model that predicts the number of Corner routes on a given play
}
\keyword{datasets}
