% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverplayersmodel}
\alias{coverplayersmodel}
\title{Cover Players Model}
\format{
A Bayesian skew_normal model
}
\usage{
coverplayersmodel
}
\description{
A model that predicts the number of defensive players in coverage on a given play
}
\keyword{datasets}
