% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_model}
\alias{cross_model}
\title{Cross Model}
\format{
A multinomial model
}
\usage{
cross_model
}
\description{
A model that predicts the number of Cross routes on a given play
}
\keyword{datasets}
