% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{def_cover_model}
\alias{def_cover_model}
\title{Defensive Coverage Snaps Model}
\format{
A linear model
}
\usage{
def_cover_model
}
\description{
A model that predicts a defensive player's coverage snaps for a given week
}
\keyword{datasets}
