% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{def_pass_rush_model}
\alias{def_pass_rush_model}
\title{Defensive Pass Rush Snaps Model}
\format{
A linear model
}
\usage{
def_pass_rush_model
}
\description{
A model that predicts a defensive player's pass rush snaps for a given week
}
\keyword{datasets}
