% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{def_run_model}
\alias{def_run_model}
\title{Run Defense Snaps Model}
\format{
A linear model
}
\usage{
def_run_model
}
\description{
A model that predicts a defensive player's run defense snaps for a given week
}
\keyword{datasets}
