% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defcoveragemodel}
\alias{defcoveragemodel}
\title{Defensive Coverage Model}
\format{
A multinomial regression model
}
\usage{
defcoveragemodel
}
\description{
A model that predicts the type of defensive coverage on a given play:
Cover 0 , Cover 1, Cover 2, Cover 3, Cover 4, Cover 6, Cover2Man, GoalLine,
Prevent, or RedZone
}
\keyword{datasets}
