% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defdf}
\alias{defdf}
\title{PFF Defense Snaps Data 2024}
\format{
A data frame with 1010 rows and 5 variables:
\describe{
\item{player}{An NFL player}
\item{position}{F7 (for front seven player) or DB (for secondary player)}
\item{mean_cov_snaps}{The mean number of coverage snaps for the player in 2024}
\item{mean_run_def_snaps}{The mean number of run defense snaps for the player in 2024}
\item{mean_pass_rush_snaps}{The mean number of pass rush snaps for the player in 2024}
}
}
\usage{
defdf
}
\description{
A data set containing NFL defensive player information on coverage snaps,
pass rush snaps, and run defense snaps.
}
\keyword{datasets}
