% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defpersonnelmodel}
\alias{defpersonnelmodel}
\title{Defensive Personnel Model}
\format{
A multinomial regression model
}
\usage{
defpersonnelmodel
}
\description{
A model that predicts the type of defensive personnel on a given play:
1-5, 2-4, 2-5, 3-3, 3-4, 4-2, 4-3, 5-1, 5-2, 6-1, Nickel, Dime, Heavy, or Goalline
}
\keyword{datasets}
