% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{depthchart}
\alias{depthchart}
\title{Returns the most updated depth chart for a selected NFL team}
\usage{
depthchart(tm)
}
\arguments{
\item{team}{A two to three letter character string abbreviation for NFL teams
Must be one of the following abbreviations:
"ARI" "ATL" "BAL" "BUF" "CAR" "CHI" "CIN" "CLE" "DAL"
"DEN" "DET" "GB"  "HOU" "IND" "JAX" "KC"  "LA"  "LAC"
"LV"  "MIA" "MIN" "NE"  "NO"  "NYG" "NYJ" "PHI" "PIT"
"SEA" "SF"  "TB"  "TEN" "WAS"}
}
\value{
A data frame representation of the depth chart with each players position
}
\description{
Returns the most updated depth chart for a selected NFL team
}
