% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nflfastR_data.R
\name{download_nflfastR_data}
\alias{download_nflfastR_data}
\title{Download raw nflfastR data in rds format}
\usage{
download_nflfastR_data(year)
}
\arguments{
\item{year}{A year from 2009 to 2020}
}
\value{
A data.frame containing play-by-play information from NFL games
}
\description{
This function will return a tibble after downloading the original file 
from the nflfastR-data website. Note that the tibble will contain all 
regular and postseason data.
}
\examples{
df <- download_nflfastR_data(2019)
}
