% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nflscrapR_data.R
\name{download_nflscrapR_data}
\alias{download_nflscrapR_data}
\title{Download raw nflscrapR data in csv format}
\usage{
download_nflscrapR_data(type = "regular", year)
}
\arguments{
\item{type}{A character string specifying "regular", "pre", or "post", for 
regular, pre, or post season, respectively.}

\item{year}{A year between 2009 and 2019}
}
\value{
A data.frame containing play-by-play information from NFL games
}
\description{
This function will return a data.frame after downloading the original file 
the nflscrapR-data website.
}
\examples{
\dontrun{
df <- download_nflscrapR_data("regular", 2019)
}
}
