% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fgmakedf}
\alias{fgmakedf}
\title{Field Goal Make Probability}
\format{
A data frame with 101 rows and 2 variables:
\describe{
\item{FGDist}{Distance of Field Goal Attempt}
\item{make_prob}{Probability of making field goal from given distance}
}
}
\usage{
fgmakedf
}
\description{
A data set containing the probability of making a field goal from distances
17-99 yards
}
\keyword{datasets}
