% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flat_model}
\alias{flat_model}
\title{Flat Model}
\format{
A multinomial model
}
\usage{
flat_model
}
\description{
A model that predicts the number of Flat routes on a given play
}
\keyword{datasets}
