% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fumblostdf}
\alias{fumblostdf}
\title{Fumble Lost Probability}
\format{
A data frame with 5 rows and 4 variables:
}
\usage{
fumblostdf
}
\description{
A data set containing the probability that a fumble was lost by a FB, QB, RB,
TE, or WR
}
\keyword{datasets}
