% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitch_model}
\alias{hitch_model}
\title{Hitch Model}
\format{
A multinomial model
}
\usage{
hitch_model
}
\description{
A model that predicts the number of Hitch routes on a given play
}
\keyword{datasets}
