% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_yds_model}
\alias{int_yds_model}
\title{Interception Yards Model}
\format{
A Bayesian skew_normal model
}
\usage{
int_yds_model
}
\description{
A model that predicts a distribution of possible interception yard values
}
\keyword{datasets}
