% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interception_model}
\alias{interception_model}
\title{Interception Model}
\format{
A logistic regression model
}
\usage{
interception_model
}
\description{
A model that predicts the probability of a passing play resulting in an interception
}
\keyword{datasets}
