% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{off_personnel_model}
\alias{off_personnel_model}
\title{Offensive Personnel Model}
\format{
A multinomial regression model
}
\usage{
off_personnel_model
}
\description{
A model that predicts the type of offensive personnel on a given play:
00, 01, 02, 03, 10, 11, 12, 13, 20, 21, 22, or 23
}
\keyword{datasets}
