% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{offdf}
\alias{offdf}
\title{All Offensive Personnel}
\format{
A data frame with 15606 rows and 296 variables:
}
\usage{
offdf
}
\description{
A data set containing offensive personnel data from the Big Data Bowl 2025
}
\keyword{datasets}
