% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{out_model}
\alias{out_model}
\title{Out Model}
\format{
A multinomial model
}
\usage{
out_model
}
\description{
A model that predicts the number of Out routes on a given play
}
\keyword{datasets}
