% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passing_yds_model}
\alias{passing_yds_model}
\title{Passing Yards Model}
\format{
A Bayesian skew_normal model
}
\usage{
passing_yds_model
}
\description{
A model that provides a distribution of possible yards on a given passing play
}
\keyword{datasets}
