% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{passresultmodel}
\alias{passresultmodel}
\title{Pass Play Result Model}
\format{
A multinomial regression model
}
\usage{
passresultmodel
}
\description{
A model that predicts the outcome of a passing play:
COMPLETION, INCOMPLETION, SACK, FUMBLE, or INTERCEPTION
}
\keyword{datasets}
