% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_def_coverage_grade_mean_model}
\alias{PFF_def_coverage_grade_mean_model}
\title{PFF Defensive Coverage Grade Mean Model}
\format{
A linear model
}
\usage{
PFF_def_coverage_grade_mean_model
}
\description{
A model that predicts the defensive coverage grade of a DB or F7 for a week
}
\keyword{datasets}
