% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_Defense_2025}
\alias{PFF_Defense_2025}
\title{PFF Defense Data 2025}
\format{
: TEMP HOLDER
}
\usage{
PFF_Defense_2025
}
\description{
A data set containing NFL player information on Defense grades from 2025
}
\keyword{datasets}
