% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_pass_blocking_grade_mean_model}
\alias{PFF_pass_blocking_grade_mean_model}
\title{PFF Pass Blocking Grade Mean Model}
\format{
A linear model
}
\usage{
PFF_pass_blocking_grade_mean_model
}
\description{
A model that predicts the pass blocking grade of a RB, WR, TE, or OL for a week
}
\keyword{datasets}
