% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_passing_grade_mean_model}
\alias{PFF_passing_grade_mean_model}
\title{PFF Passing Grade Mean Model}
\format{
A linear model
}
\usage{
PFF_passing_grade_mean_model
}
\description{
A model that predicts the passing grade of an NFL Quarterback for a week
}
\keyword{datasets}
