% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_QBs_2024}
\alias{PFF_QBs_2024}
\title{PFF Passing Data 2024}
\format{
: TEMP HOLDER
}
\usage{
PFF_QBs_2024
}
\description{
A data set containing NFL player information on QB grades from 2024
}
\keyword{datasets}
