% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_Receiving_2024}
\alias{PFF_Receiving_2024}
\title{PFF Receiving Data 2024}
\format{
: TEMP HOLDER
}
\usage{
PFF_Receiving_2024
}
\description{
A data set containing NFL player information on Receiving grades from 2024
}
\keyword{datasets}
