% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_receiving_grade_mean_model}
\alias{PFF_receiving_grade_mean_model}
\title{PFF Receiving Grade Mean Model}
\format{
A linear model
}
\usage{
PFF_receiving_grade_mean_model
}
\description{
A model that predicts the receiving grade of an NFL RB, WR, or TE for a week
}
\keyword{datasets}
