% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFF_rushing_grade_mean_model}
\alias{PFF_rushing_grade_mean_model}
\title{PFF Rushing Grade Mean Model}
\format{
A linear model
}
\usage{
PFF_rushing_grade_mean_model
}
\description{
A model that predicts the rushing grade of an NFL Runningback for a week
}
\keyword{datasets}
