% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posroutedf}
\alias{posroutedf}
\title{Position Route Run Probability Data}
\format{
A data frame:
}
\usage{
posroutedf
}
\description{
A data set containing probability that a route was run by a certain position.
}
\keyword{datasets}
