% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{post_model}
\alias{post_model}
\title{Post Model}
\format{
A multinomial model
}
\usage{
post_model
}
\description{
A model that predicts the number of Post routes on a given play
}
\keyword{datasets}
