% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pbp_data.R
\name{prep_pbp_data}
\alias{prep_pbp_data}
\title{Add necessary columns to nflscrapR data}
\usage{
prep_pbp_data(data)
}
\arguments{
\item{data}{An nflscrapR or nflfastR data set. Note that stringsAsFactors = FALSE is assumed.}
}
\value{
a data.table object
}
\description{
Add necessary columns to nflscrapR data
}
\examples{
\dontrun{
dt <- prep_pbp_data(nflscrapr_pbp_data)
}
}
