% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressureplayersmodel}
\alias{pressureplayersmodel}
\title{Pressure Players Model}
\format{
An ordinal regression model
}
\usage{
pressureplayersmodel
}
\description{
A model that predicts the number of players putting pressure on the QB on
a given play
}
\keyword{datasets}
