% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qbsnapsdf2024}
\alias{qbsnapsdf2024}
\title{PFF QB Rushing Snaps Data 2024}
\format{
A data frame

}
}
\usage{
qbsnapsdf2024
}
\description{
A data set containing mean number of rushing attempts for NFL QBs in 2024.
}
\keyword{datasets}
