% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec_route_model}
\alias{rec_route_model}
\title{Receiver Route Model}
\format{
A linear model
}
\usage{
rec_route_model
}
\description{
A model that predicts a recievers (WR or TE) routes ran for a given week
}
\keyword{datasets}
