% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recdf}
\alias{recdf}
\title{PFF Receiving Snaps Data 2024}
\format{
A data frame with 390 rows and 3 variables:
\describe{
\item{player}{An NFL player}
\item{position}{WR (for wide receiver) or TE (for tight end)}
\item{mean_snaps}{The mean number of routes run for the player in 2024}
}
}
\usage{
recdf
}
\description{
A data set containing mean number of routes run for NFL WRs and TEs in 2024.
}
\keyword{datasets}
