% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_pbp_2018}
\alias{reg_pbp_2018}
\title{NFL Play-by-Play Data 2018}
\format{
A data frame with 45156 rows and 256 variables:
\describe{
  \item{play_id}{id within a game}
  \item{game_id}{id of the specific game}
  \item{home_team}{home team playing}
  \item{away_team}{away team playing}
  \item{posteam}{team with possession}
  ...
}
}
\source{
\url{https://ryurko.github.io/nflscrapR-data/}
}
\usage{
reg_pbp_2018
}
\description{
A dataset containing the 256 attributes associated with 45156 plays in the
NFL from the 2018 regular season. The data were obtained using the nflscrapR package.
}
\keyword{datasets}
