% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_count_model}
\alias{route_count_model}
\title{Route Count Model}
\format{
A multinomial model
}
\usage{
route_count_model
}
\description{
A model that predicts the number of routes on a given play
}
\keyword{datasets}
