% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runpass_model}
\alias{runpass_model}
\title{Run or Pass Model}
\format{
A logistic regression model
}
\usage{
runpass_model
}
\description{
A model that predicts whether a given play will be a run or a pass
}
\keyword{datasets}
