% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rush_att_model}
\alias{rush_att_model}
\title{Rush Attempts Model}
\format{
A linear model
}
\usage{
rush_att_model
}
\description{
A model that predicts a running back's rush attempts for a given week
}
\keyword{datasets}
