% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rushdf}
\alias{rushdf}
\title{PFF Rushing Snaps Data 2024}
\format{
A data frame with 138 rows and 3 variables:
\describe{
\item{player}{An NFL player}
\item{position}{All RB (for running back)}
\item{mean_snaps}{The mean number of rush attempts for the player in 2024}
}
}
\usage{
rushdf
}
\description{
A data set containing mean number of rushing attempts for NFL RBs in 2024.
}
\keyword{datasets}
