% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rushfumble_model}
\alias{rushfumble_model}
\title{Rushing Play Fumble Model}
\format{
A logistic regression model
}
\usage{
rushfumble_model
}
\description{
A model that predicts the probability that a rushing play will result in a
fumble
}
\keyword{datasets}
