% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rushing_yds_model}
\alias{rushing_yds_model}
\title{Rushing Yards Model}
\format{
A Bayesian skew_normal model
}
\usage{
rushing_yds_model
}
\description{
A model that provides a distribution for rushing yards on a given play
}
\keyword{datasets}
