% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sack_model}
\alias{sack_model}
\title{Sack Model}
\format{
A logistic regression model
}
\usage{
sack_model
}
\description{
The probability that a passing play results in a sack
}
\keyword{datasets}
