% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sack_yards_model}
\alias{sack_yards_model}
\title{Sack Yards Model}
\format{
A Bayesian gaussian model
}
\usage{
sack_yards_model
}
\description{
A model that provides a distribution for sack yards on a given play that
resulted in a sack
}
\keyword{datasets}
