% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_drives.R
\name{sample_drives}
\alias{sample_drives}
\title{Sample a Series of Drives, a strategy to test verses the normal strategy}
\usage{
sample_drives(
  n_sims,
  from_yard_line = 25,
  play_by_play_data,
  strategy = "normal",
  single_drive = FALSE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{n_sims}{The number of simulations}

\item{from_yard_line}{The starting field position (defaults to 25)}

\item{play_by_play_data}{A data file from nflscrapR prepped using the prep_pbp_data.R function}

\item{strategy}{"normal", "passes_rushes", or "fourth_downs"}

\item{single_drive}{TRUE indicates only a single drive, otherwise, drives sampled until a score occurs}

\item{progress}{logical for inclusion of a progress bar}

\item{...}{Additional arguments for different strategies}
}
\value{
A data.frame of drives
}
\description{
Sample a Series of Drives, a strategy to test verses the normal strategy
}
\examples{
\dontrun{
sample_drives(2, 25, dt)
}

}
