% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_drives_until_score.R
\name{sample_drives_until_score}
\alias{sample_drives_until_score}
\title{Sample a Series of Drives Until a Score, a strategy to test verses the normal strategy}
\usage{
sample_drives_until_score(
  n_sims,
  from_yard_line = 25,
  play_by_play_data,
  strategy1 = "normal",
  progress = TRUE,
  single_drive = F,
  ...
)
}
\arguments{
\item{n_sims}{The number of simulations}

\item{from_yard_line}{The starting field position (defaults to 25)}

\item{play_by_play_data}{A data file from nflscrapR prepped using the prep_pbp_data.R function}

\item{progress}{logical for inclusion of a progress bar}

\item{STRATEGY1}{"normal", "passes_rushes", or "fourth_downs"}
}
\value{
A data.frame of drives
}
\description{
Sample a Series of Drives Until a Score, a strategy to test verses the normal strategy
}
\examples{
\dontrun{
sample_drives_until_score(2, 25, dt,
                          STRATEGY1 = sample_play,
                          STRATEGY2 = sample_play_go_for_it)
}

}
