% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_play_test.R
\name{sample_play_test}
\alias{sample_play_test}
\title{Sample NFL Play-by-Play Data}
\usage{
sample_play_test(what_down, yards_to_go, yards_from_own_goal,
  play_by_play_data, strategy = "normal", prop_passes = NULL,
  strat = "empirical", yards_less_than = 5)
}
\arguments{
\item{what_down}{The current down (1st, 2nd, 3rd, or 4th down)}

\item{yards_to_go}{Number of yards to go until a first down or TD}

\item{yards_from_own_goal}{The number of yards from the possession team's own goal}

\item{play_by_play_data}{A data file from nflscrapR prepped using the prep_pbp_data.R function#}

\item{passes_and_rushes}{Logical parameter, TRUE if this function should sample both passing and rushing plays}

\item{just_passes}{Logical parameter, TRUE if this function should only sample passing plays}

\item{just_rushes}{Logical parameter, TRUE if this function should only sample rushing plays}
}
\value{
A tibble containing lots of info
}
\description{
This function will return a sample play from the nflscrapR play-by-play
data for a given down, distance, yards from the team's goal, using the
usual NFL-coaching strategy.
}
\examples{
\dontrun{
sample_play_test(what_down = 3,
            yards_to_go = 2,
            yards_from_own_goal = 45,
            play_by_play_data = pbp_data,
            strategy = "normal")
}
}
