% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{screen_model}
\alias{screen_model}
\title{Screen Model}
\format{
A multinomial model
}
\usage{
screen_model
}
\description{
A model that predicts the number of Screen routes on a given play
}
\keyword{datasets}
