% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slant_model}
\alias{slant_model}
\title{Slant Model}
\format{
A multinomial model
}
\usage{
slant_model
}
\description{
A model that predicts the number of Slant routes on a given play
}
\keyword{datasets}
