% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatorFunctions.R
\name{snaps_to_depth_chart}
\alias{snaps_to_depth_chart}
\title{Returns the most updated ESPN depth chart for a selected NFL team with
projected snaps for 2025}
\usage{
snaps_to_depth_chart(team, def_assignment)
}
\arguments{
\item{team}{A two to three letter character string abbreviation for NFL teams
Must be one of the following abbreviations:
"ARI" "ATL" "BAL" "BUF" "CAR" "CHI" "CIN" "CLE" "DAL"
"DEN" "DET" "GB"  "HOU" "IND" "JAX" "KC"  "LA"  "LAC"
"LV"  "MIA" "MIN" "NE"  "NO"  "NYG" "NYJ" "PHI" "PIT"
"SEA" "SF"  "TB"  "TEN" "WAS"}

\item{def_assignment}{A string of what type of defensive assignment you want
snaps for
Must be one of the following:
"coverage", "run_defense", "pass_rush"}
}
\value{
A data frame representation of the depth chart with each players position
}
\description{
Returns the most updated ESPN depth chart for a selected NFL team with
projected snaps for 2025
}
