% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetted_route_model}
\alias{targetted_route_model}
\title{Targetted Route Model}
\format{
A multinomial regression model
}
\usage{
targetted_route_model
}
\description{
A model that predicts the route which will be targetted on a given play
}
\keyword{datasets}
