% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teamrushposdf}
\alias{teamrushposdf}
\title{Team Position Rushing Probabilites}
\format{
A dataframe with 113 rows and 4 variables
}
\usage{
teamrushposdf
}
\description{
A data set containing the probability of a certain position rushing the ball
for each NFL team
}
\keyword{datasets}
