% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeToThrowmodel}
\alias{timeToThrowmodel}
\title{Time to Throw Model}
\format{
A Bayesian skew_normal model
}
\usage{
timeToThrowmodel
}
\description{
A model that provides a distribution for the time a quarterback will have
to throw the ball on a given play.
}
\keyword{datasets}
