% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheel_model}
\alias{wheel_model}
\title{Wheel Model}
\format{
A multinomial model
}
\usage{
wheel_model
}
\description{
A model that predicts the number of Wheel routes on a given play
}
\keyword{datasets}
