% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_drives_until_score.R
\name{sample_drives_until_score}
\alias{sample_drives_until_score}
\title{Sample a Series of Drives Until a Score}
\usage{
sample_drives_until_score(n_sims, from_yard_line = 25, play_by_play_data,
  STRATEGY1 = sample_play, STRATEGY2 = sample_play, progress = TRUE)
}
\arguments{
\item{n_sims}{The number of simulations}

\item{from_yard_line}{The starting field position (defaults to 25)}

\item{play_by_play_data}{A data file from nflscrapR prepped using the prep_pbp_data.R function}

\item{STRATEGY1}{one of sample_play or sample_play_go_for_it}

\item{STRATEGY2}{one of sample_play or sample_play_go_for_it}

\item{progress}{logical for inclusion of a progress bar}
}
\value{
A data.frame of drives
}
\description{
Sample a Series of Drives Until a Score
}
\examples{
\dontrun{
sample_drives_until_score(2, 25, dt,
                          STRATEGY1 = sample_play,
                          STRATEGY2 = sample_play_go_for_it)
}

}
