% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_play.R
\name{sample_play}
\alias{sample_play}
\title{Sample NFL Play-by-Play Data}
\usage{
sample_play(what_down, yards_to_go, yards_from_own_goal, play_by_play_data)
}
\arguments{
\item{what_down}{The current down (1st, 2nd, 3rd, or 4th down)}

\item{yards_to_go}{Number of yards to go until a first down or TD}

\item{yards_from_own_goal}{The number of yards from the possession team's own goal}

\item{play_by_play_data}{A data file from nflscrapR prepped using the prep_pbp_data.R function#}
}
\value{
A tibble containing lots of info
}
\description{
This function will return a sample play from the nflscrapR play-by-play
data for a given down, distance, yards from the team's goal, using the
usual NFL-coaching strategy.
}
\examples{
\dontrun{
sample_play(what_down = 3,
            yards_to_go = 2,
            yards_from_own_goal = 45,
            play_by_play_data = pbp_data)
}
}
